Acute Risks in Micromorts

A comprehensive dataset of activities and their associated acute mortality risk in micromorts, with calculated microlives and source references.

Usage

common_risks()

Value

A tibble with columns: activity, micromorts, microlives, category, period, source_url.

Details

Micromort: one-in-a-million chance of death (acute risk). Microlife: 30 minutes of life expectancy lost.

Data sources: Wikipedia, micromorts.rip, CDC MMWR, academic literature.

References

Howard RA (1980). "On Making Life and Death Decisions." In Schwing & Albers (eds), Societal Risk Assessment: How Safe Is Safe Enough?

https://en.wikipedia.org/wiki/Micromort

https://micromorts.rip/

Examples

common_risks()
#> # A tibble: 62 × 6
#>    activity                     micromorts microlives category period source_url
#>    <chr>                             <dbl>      <dbl> <chr>    <chr>  <chr>     
#>  1 Mt. Everest ascent                37932     26552. Mountai… per a… https://m…
#>  2 Himalayan mountaineering          12000      8400  Mountai… per e… https://m…
#>  3 COVID-19 infection (unvacci…      10000      7000  Disease  per i… https://m…
#>  4 Spanish flu infection              3000      2100  Disease  per i… https://m…
#>  5 Matterhorn ascent                  2840      1988  Mountai… per a… https://m…
#>  6 Living in US during COVID-1…        500       350  Disease  per m… https://m…
#>  7 Living (one day, age 90)            463       324. Daily L… per d… https://m…
#>  8 Base jumping (per jump)             430       301  Sport    per e… https://m…
#>  9 First day of life (newborn)         430       301  Daily L… per d… https://m…
#> 10 COVID-19 unvaccinated (age …        234       164. COVID-19 11 we… https://w…
#> # ℹ 52 more rows
common_risks() |> dplyr::filter(category == "COVID-19")
#> # A tibble: 15 × 6
#>    activity                     micromorts microlives category period source_url
#>    <chr>                             <dbl>      <dbl> <chr>    <chr>  <chr>     
#>  1 COVID-19 unvaccinated (age …     234         164.  COVID-19 11 we… https://w…
#>  2 COVID-19 unvaccinated (age …      76          53.2 COVID-19 11 we… https://w…
#>  3 COVID-19 monovalent vaccine…      55          38.5 COVID-19 11 we… https://w…
#>  4 COVID-19 bivalent booster (…      23          16.1 COVID-19 11 we… https://w…
#>  5 COVID-19 unvaccinated (all …      20          14   COVID-19 11 we… https://w…
#>  6 COVID-19 monovalent vaccine…       9           6.3 COVID-19 11 we… https://w…
#>  7 COVID-19 unvaccinated (age …       8           5.6 COVID-19 11 we… https://w…
#>  8 COVID-19 monovalent vaccine…       4           2.8 COVID-19 11 we… https://w…
#>  9 COVID-19 bivalent booster (…       3           2.1 COVID-19 11 we… https://w…
#> 10 COVID-19 monovalent vaccine…       2           1.4 COVID-19 11 we… https://w…
#> 11 COVID-19 unvaccinated (age …       1           0.7 COVID-19 11 we… https://w…
#> 12 COVID-19 bivalent booster (…       1           0.7 COVID-19 11 we… https://w…
#> 13 COVID-19 bivalent booster (…       1           0.7 COVID-19 11 we… https://w…
#> 14 COVID-19 monovalent vaccine…       0.2         0.1 COVID-19 11 we… https://w…
#> 15 COVID-19 bivalent booster (…       0.05        0   COVID-19 11 we… https://w…
common_risks() |> dplyr::filter(micromorts > 100)
#> # A tibble: 14 × 6
#>    activity                     micromorts microlives category period source_url
#>    <chr>                             <dbl>      <dbl> <chr>    <chr>  <chr>     
#>  1 Mt. Everest ascent                37932    26552.  Mountai… per a… https://m…
#>  2 Himalayan mountaineering          12000     8400   Mountai… per e… https://m…
#>  3 COVID-19 infection (unvacci…      10000     7000   Disease  per i… https://m…
#>  4 Spanish flu infection              3000     2100   Disease  per i… https://m…
#>  5 Matterhorn ascent                  2840     1988   Mountai… per a… https://m…
#>  6 Living in US during COVID-1…        500      350   Disease  per m… https://m…
#>  7 Living (one day, age 90)            463      324.  Daily L… per d… https://m…
#>  8 Base jumping (per jump)             430      301   Sport    per e… https://m…
#>  9 First day of life (newborn)         430      301   Daily L… per d… https://m…
#> 10 COVID-19 unvaccinated (age …        234      164.  COVID-19 11 we… https://w…
#> 11 Caesarean birth (mother)            170      119   Medical  per e… https://m…
#> 12 Scuba diving (per year, tra…        164      115.  Sport    per y… https://m…
#> 13 Vaginal birth (mother)              120       84   Medical  per e… https://m…
#> 14 Living (one day, age 75)            105       73.5 Daily L… per d… https://m…